+++ gui/tabs/tabsimulate.cpp
@@ -381,8 +381,16 @@ void TabSimulate::btnSimulate_clicked() {
 }
 
 void TabSimulate::btnSliceImage_clicked() {
+  DialogImageSlicePlotData plotData;
+  plotData.simulation = sim.get();
+  plotData.compartmentNames = compartmentNames;
+  plotData.speciesToDraw = compartmentSpeciesToDraw;
+  plotData.timeUnit = model.getUnits().getTime().name;
+  plotData.lengthUnit = model.getUnits().getLength().name;
+  plotData.concentrationUnit = model.getUnits().getConcentration();
+  plotData.timepointIndex = ui->hslideTime->value();
   DialogImageSlice dialog(model.getGeometry().getImages(), images, time,
-                          flipYAxis);
+                          flipYAxis, plotData);
   if (dialog.exec() == QDialog::Accepted) {
     SPDLOG_DEBUG("todo: save current slice settings");
   }